Add IActionResult response in Controllers (#4)

+++ RentACarWebAPI/Controllers/CarController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
 using RentACarWebAPI.DTOs;
 using RentACarWebAPI.Interfaces.Services;
 using System.Collections.Generic;
@@ -21,57 +22,85 @@ namespace RentACarWebAPI.Controllers
 
         // GET: api/<CarController>
         [HttpGet]
-        public IEnumerable<CarDto> Get()
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IEnumerable<CarDto>))]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        public IActionResult Get()
         {
             var carEntityList = _carService.GetAll();
 
-            var carDtoList = carEntityList.Select(carEntity => new CarDto(carEntity)).ToList();
+            if (carEntityList is null)
+                return NotFound();
 
-            return carDtoList;
+            var carDtoList = carEntityList.Select(carEntity => new CarDto(carEntity));
+
+            return Ok(carDtoList);
         }
 
         // GET api/<CarController>/5
         [HttpGet("{id}")]
-        public CarDto Get(int id)
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(CarDto))]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        public IActionResult Get(int id)
         {
             var carEntity = _carService.Get(id);
 
+            if (carEntity is null)
+                return NotFound();
+
             var carDto = new CarDto(carEntity);
 
-            return carDto;
+            return Ok(carDto);
         }
 
         // POST api/<CarController>
         [HttpPost]
-        public CarDto Post([FromBody] CarDto carDto)
+        [ProducesResponseType(StatusCodes.Status201Created, Type = typeof(CarDto))]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        public IActionResult Post([FromBody] CarDto carDto)
         {
             var carEntity = carDto.ToCarEntity(carDto);
 
             var carCreated = _carService.Create(carEntity);
 
+            if (carCreated is null)
+                return NotFound();
+
             var carResponse = new CarDto(carCreated);
 
-            return carResponse;
+            return Ok(carResponse);
         }
 
         // PUT api/<CarController>/5
         [HttpPut("{id}")]
-        public CarDto Put([FromBody] CarDto carDto, int id)
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(CarDto))]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        public IActionResult Put([FromBody] CarDto carDto, int id)
         {
             var carEntity = carDto.ToCarEntity(carDto);
 
-            var carCreated = _carService.Update(carEntity, id);
+            var carUpdated = _carService.Update(carEntity, id);
 
-            var carResponse = new CarDto(carCreated);
+            if (carUpdated is null)
+                return NotFound();
 
-            return carResponse;
+            var carResponse = new CarDto(carUpdated);
+
+            return Ok(carResponse);
         }
 
         // DELETE api/<CarController>/5
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public IActionResult Delete(int id)
         {
+            var carExist = _carService.EntityExist(id);
+
+            if (!carExist)
+                return NotFound();
+
             _carService.Delete(id);
+            return NoContent();
         }
     }
 }
